move to_activity to Base class
